fixed the ml db info getting pushed through to the frontend

+++ backend/IntexAPI/IntexAPI/Data/DonorRiskScore.cs
@@ -8,7 +8,7 @@ public class DonorRiskScore
 {
     [Key]
     [Column("supporter_id")]
-    public int SupporterId { get; set; }
+    public long SupporterId { get; set; }
 
     [Column("display_name")]
     public string? DisplayName { get; set; }
@@ -25,13 +25,13 @@ public class DonorRiskScore
     public string? RiskLabel { get; set; }
 
     [Column("predicted_at_risk")]
-    public int? PredictedAtRisk { get; set; }
+    public long? PredictedAtRisk { get; set; }
 
     [Column("recency_days")]
-    public int? RecencyDays { get; set; }
+    public long? RecencyDays { get; set; }
 
     [Column("frequency")]
-    public int? Frequency { get; set; }
+    public long? Frequency { get; set; }
 
     [Column("prediction_timestamp")]
     public string? PredictionTimestamp { get; set; }
